lune/dev: explicitly bind to IPv4 loopback
Currently the HTTP server is started with an explicit host and the WebSocket Server is left to pick its own. This can result in IPv6 being preferred on the WebSocket server side when the WebSocket client points to IPv4 loopback explicitly.

This change fixes that by explicitly specifying the IPv4 loopback address.